Move team password command out of `/team`
This allows its permissions to be set independently of `/team`.
